spm: use virtual counter for sp_sleep
SPs(at S-EL1) can only access virtual counters which is calculated as:
  Virtual Count = Physical Count - <offset>
Offset value is specified in the register CNTVOFF_EL2, which is set
as 0 by hafnium.

This patch changes sp_sleep() to use virtual counter instead of
physical counter along with changing its argument from uint32_t
to uint16_t as required by next change.

+++ spm/common/sp_helpers.h
@@ -66,7 +66,7 @@ void announce_test_start(const char *test_desc);
 void announce_test_end(const char *test_desc);
 
 /* Sleep for at least 'ms' milliseconds. */
-void sp_sleep(uint32_t ms);
+void sp_sleep(uint16_t ms);
 
 /*
  * Hypervisor Calls Wrappers
